% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.nnmi.R
\name{nnmi}
\alias{nnmi}
\title{A 'nnmi' class}
\usage{
nnmi(dat, ...)
}
\arguments{
\item{dat}{a list of imputed data set.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A list-based S3 class for storing imputation results from nearest neighbor based multiple imputation
}
\examples{
# load required packages
library(NNMIS)
library(survival)

# load data set - stanford2 in package 'survival'
data("stanford2")
head(stanford2)
attach(stanford2)

# performance multiple imputation on missing covariate t5
imp.dat <- NNMIS(t5, xa=age, xb=age, time=time, event=status, Seed = 1234, mc.core=1)

# S3 method for class 'nnmi'
print(imp.dat)


}
