% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKM.R
\name{plotKM}
\alias{plotKM}
\title{Plot function for pooled Kaplan-Meier estimates}
\usage{
plotKM(x)
}
\arguments{
\item{x}{a data.frame contains pooled estimates of survival function generated from function 'km.pool'.}
}
\description{
A plot of survival curves is produced.
}
\seealso{
km.pool
}
