% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nnmi.R
\name{print.nnmi}
\alias{print.nnmi}
\title{Print function for object of 'nnmi' class.}
\usage{
\method{print}{nnmi}(x, ...)
}
\arguments{
\item{x}{a 'nnmi' object}

\item{...}{further arguments passed to function}
}
\description{
Print function for object of 'nnmi' class.
}
