% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nnmi.R
\name{print.nnmi}
\alias{print.nnmi}
\title{Print function for object of 'nnmi' class.}
\usage{
\method{print}{nnmi}(x, ...)
}
\arguments{
\item{x}{a 'nnmi' object}

\item{...}{further arguments passed to function}
}
\description{
print basic information for an object of class 'nnmi'
}
\examples{
# load required packages
library(NNMIS)
library(survival)

# load data set - stanford2 in package 'survival'
data("stanford2")
head(stanford2)
attach(stanford2)

# performance multiple imputation on missing covariate t5
imp.dat <- NNMIS(t5, xa=age, xb=age, time=time, event=status, Seed = 1234, mc.core=1)

# print
imp.dat

}
